pyproject.toml
[project]
name = "color-mixer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "coloraide-extras>=1.11",
    "coloraide[all]>=4.7.2",
    "colour-science>=0.4.6",
    "flask>=3.1.1",
    "matplotlib>=3.10.5",
    "numpy>=2.3.2",
    "pip>=25.2",
    "pytest>=8.4.1",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
